Better structure for native source Ignore built binaries
