+++ docker.sh
@@ -1,3 +1,5 @@
-cd /code/src/lib/
+set -e
+
+cd /code/
 make windows
 make linux
